(feat): Rename data_structures package
